feat(configureStoreDev): add dev store
